+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Bug Fixes:
     * examples: tetris code cleanup, use custom fonts
     * compiler: fix float handling in DATA statements
     * compiler: fix coord/coord2 error handling
+    * compiler: do not abort on type system inconsistencies (e.g. unresolved forwarded types)
     * use EXIT\_FAILURE for fatal error conditions (fixes issue #13 by polluks)
     * add "$VER" version string